refactor: 경로, 파일명, 폴더 구조 수정
