--- layouts/Footer.tsx
@@ -1,8 +1,0 @@
-/** 푸터 */
-const Footer = () => {
-  return (
-    <></>
-  );
-};
-
-export default Footer;
