--- layouts/Header.tsx
@@ -1,8 +1,0 @@
-/** 헤더 */
-const Header = () => {
-  return (
-    <></>
-  );
-};
-
-export default Header;
