+++ pages/_app.tsx
@@ -1,34 +1,25 @@
-import { Inter } from 'next/font/google';
-import type { AppProps } from 'next/app';
-import AppLayout from 'components/common/AppLayout';
-import Globals from 'styles/Globals';
-import Login from './login/Login';
+import { Inter } from "next/font/google";
+import type { AppProps } from "next/app";
+import AppLayout from "layouts/AppLayout";
+import Globals from "styles/Globals";
+import Login from "./login";
 
-const inter = Inter({ subsets: ['latin'] });
- 
+const inter = Inter({ subsets: ["latin"] });
 
 const App = ({ Component, pageProps }: AppProps) => {
-  if (Component === Login) {
-    return (
-      <main className={inter.className}>
-        <Component {...pageProps}/>
-        <Globals />
-      </main>
-    );
-  }
-
-  return (
+  return Component === Login ? (
+    <main className={inter.className}>
+      <Component {...pageProps} />
+      <Globals />
+    </main>
+  ) : (
     <main className={inter.className}>
       <AppLayout>
-        <Component {...pageProps}/>
+        <Component {...pageProps} />
         <Globals />
       </AppLayout>
-     
     </main>
-  )
-}
+  );
+};
 
- 
 export default App;
-
-
