--- pages/community/Community.tsx
@@ -1,12 +1,0 @@
-import Header from "../../components/common/header";
-
-const Community  = () => {
-    return (
-      <div>
-      <h1>community page</h1>
-      {/* Add your content here */}
-      </div>
-    );
-  };
-  
-  export default Community;
