+++ pages/community/index.tsx
@@ -1,0 +1,12 @@
+import Header from "../../components/common/Header/header";
+
+const Community = () => {
+  return (
+    <div>
+      <h1>community page</h1>
+      {/* Add your content here */}
+    </div>
+  );
+};
+
+export default Community;
