--- pages/connect/Connect.tsx
@@ -1,12 +1,0 @@
-import Header from "../../components/common/header";
-
-const Connect  = () => {
-  return (
-    <div>
-      <h1>connect page</h1>
-      {/* Add your content here */}
-    </div>
-  );
-};
-
-export default Connect;
