+++ pages/connect/index.tsx
@@ -1,0 +1,12 @@
+import Header from "../../components/common/Header/header";
+
+const Connect = () => {
+  return (
+    <div>
+      <h1>connect page</h1>
+      {/* Add your content here */}
+    </div>
+  );
+};
+
+export default Connect;
