+++ pages/index.tsx
@@ -1,11 +1,9 @@
-// import HeadPart from './HeadPart';
-import Header from '../components/common/header';
 const Home = () => {
   return (
     <>
-      <h1>hi</h1>
+      <h1>Home Page</h1>
     </>
-  )
-}
+  );
+};
 
 export default Home;
