+++ pages/myPage/index.tsx
@@ -1,0 +1,12 @@
+import Header from "../../components/common/Header/header";
+
+const MyPage = () => {
+  return (
+    <div>
+      <h1>MyPage</h1>
+      {/* Add your content here */}
+    </div>
+  );
+};
+
+export default MyPage;
