--- pages/mypage/Mypage.tsx
@@ -1,12 +1,0 @@
-import Header from "../../components/common/header";
-
-const MyPage  = () => {
-  return (
-    <div>
-      <h1>MyPage</h1>
-      {/* Add your content here */}
-    </div>
-  );
-};
-
-export default MyPage;
